[web\schema] merge schema into module
